Abstract the updating and creating of CIVSets
