+++ app/grandchallenge/reader_studies/views.py
@@ -19,7 +19,7 @@ from django.http import (
     JsonResponse,
 )
 from django.shortcuts import get_object_or_404
-from django.utils.functional import cached_property, empty
+from django.utils.functional import cached_property
 from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 from django.utils.timezone import now
@@ -59,9 +59,10 @@ from grandchallenge.components.serializers import (
 )
 from grandchallenge.components.views import (
     CIVSetBulkDeleteView,
+    CIVSetCreateViewMixin,
     CIVSetDeleteView,
-    CIVSetFormMixin,
     CivSetListView,
+    CIVSetUpdateViewMixin,
     FileUpdateBaseView,
     InterfacesCreateBaseView,
     MultipleCIVProcessingBaseView,
@@ -1261,7 +1262,9 @@ class QuestionWidgetsView(BaseAddObjectToReaderStudyMixin, View):
         return HttpResponse(form["widget"])
 
 
-class DisplaySetUpdateView(CIVSetFormMixin, MultipleCIVProcessingBaseView):
+class DisplaySetUpdateView(
+    CIVSetUpdateViewMixin, MultipleCIVProcessingBaseView
+):
     form_class = DisplaySetUpdateForm
     permission_required = (
         f"{ReaderStudy._meta.app_label}.change_{DisplaySet._meta.model_name}"
@@ -1307,28 +1310,6 @@ class DisplaySetUpdateView(CIVSetFormMixin, MultipleCIVProcessingBaseView):
             "reader-studies:display-set-interfaces-create",
             kwargs={"slug": self.base_object.slug, "pk": self.object.pk},
         )
-
-    def process_data_for_object(self, data):
-        """Updates the display set"""
-        instance = self.object
-        non_civ_data_keys = self.form_class.Meta.non_civ_fields
-
-        save = False
-        for key in non_civ_data_keys:
-            value = data.get(key, empty)
-            if value is not empty and value != getattr(instance, key):
-                setattr(instance, key, value)
-                save = True
-        if save:
-            instance.save()
-        for key in data:
-            if key in non_civ_data_keys:
-                continue
-            instance.create_civ(
-                ci_slug=key, new_value=data[key], user=self.request.user
-            )
-
-        return instance
 
 
 class DisplaySetFilesUpdate(FileUpdateBaseView):
@@ -1386,7 +1367,7 @@ class DisplaySetInterfacesCreate(InterfacesCreateBaseView):
 
 
 class DisplaySetCreateView(
-    CIVSetFormMixin,
+    CIVSetCreateViewMixin,
     MultipleCIVProcessingBaseView,
 ):
     form_class = DisplaySetCreateForm
@@ -1427,23 +1408,10 @@ class DisplaySetCreateView(
             kwargs={"slug": self.base_object.slug},
         )
 
-    def process_data_for_object(self, data):
-        """Creates a display set"""
-        non_civ_data_keys = self.form_class.Meta.non_civ_fields
-
-        instance = DisplaySet.objects.create(
-            reader_study=self.base_object,
-            **{k: data.get(k) for k in non_civ_data_keys},
+    def create_object(self, non_civ_data):
+        return DisplaySet.objects.create(
+            reader_study=self.base_object, **non_civ_data
         )
-
-        for key in data:
-            if key in non_civ_data_keys:
-                continue
-            instance.create_civ(
-                ci_slug=key, new_value=data[key], user=self.request.user
-            )
-
-        return instance
 
     def get_success_url(self):
         return self.return_url
